[Spring 장바구니 - 2단계] 소주캉(강승주) 미션 제출합니다. (#109)
* style: 개행 제거

* refactor: Customer 속성인 Email, Username, Password 값 포장

* refactor: isSameUsername 메서드를 통해 비교하도록 변경

* refactor: rename TokenRequest -> LoginRequest, TokenResponse -> LoginTokenResponse

* test: Customer 인수 테스트, Controller 테스트 분리

* feat: Product Quantity 속성 추가

* refactor: customerName -> customerId 이용하도록 API 에 맞게 수정

* feat: 장바구니 상품 목록 조회, 상품 추가 기능에 quantity, count 추가

* feat: 장바구니 상품 삭제 API 구현

* feat: 장바구니 상품 구매 개수 수정 API 구현

* feat: 장바구니 상품 추가 예외처리 구현

* feat: Cart 관련 요청 인가 기능 추가

* refactor: ObjectMapper 용 기본 생성자 접근자 public -> private

* refactor: CORS 처리 Filter 로 변경

* refactor: parameter final 제거

* refactor: 패키지 분리

* fix: 회원가입, 로그인 인증 제외

* fix: 내 장바구니 조회시 재고, 구매 수량 미포함 수정

* feat: 회원 탈퇴시 비밀번호로 검증하는 기능 추가

* refactor: 인가 응답에 password 포함하지 않도록 변경

* feat: CartItem 값 객체 분리

* sql: 회원 탈퇴시 장바구니 삭제되도록 변경

* test: message 검증 추가

* refactor: test, prod 설정 파일, schema 분리

* refactor: Product 필드 값 포장, 삭제 실패시 예외 처리, 테스트 추가

* test: CartService 에 관한 test 추가

* refactor: CartItem 관련 요청 NotNull 검증 추가

* refactor: CartService -> Cart 비즈니스 로직 이동

* feat: 주문 기능 추가

* feat: 주문 전체 조회 기능 추가

* feat: 단일 주문 조회 기능 추가

* feat: production schema 주문 관련 테이블 추가

* feat: 토큰과 요청 URL path 상의 Id를 검증하는 기능 추가ㅣ

* refactor: 필드 명 변경 value -> {element}s

* refactor: 주문 조회시 OrderId를 포함하도록 변경

* refactor: 주문 조회시 OrderId를 포함하도록 변경

+++ src/main/java/woowacourse/shoppingcart/application/CustomerService.java
@@ -1,6 +1,7 @@
 package woowacourse.shoppingcart.application;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import woowacourse.shoppingcart.dao.CustomerDao;
 import woowacourse.shoppingcart.domain.customer.Customer;
@@ -10,6 +11,7 @@ import woowacourse.shoppingcart.exception.DuplicateCustomerException;
 import woowacourse.shoppingcart.exception.InvalidCustomerException;
 
 @Service
+@Transactional(rollbackFor = Exception.class)
 public class CustomerService {
 
     private final CustomerDao customerDao;
@@ -54,7 +56,7 @@ public class CustomerService {
             .orElseThrow(InvalidCustomerException::new);
     }
 
-    public void update(Long id, CustomerUpdateRequest request) {
+    public void update(long id, CustomerUpdateRequest request) {
         if (isSameOriginUsername(id, request)) {
             return;
         }
@@ -63,17 +65,20 @@ public class CustomerService {
         customerDao.update(id, request.getUsername());
     }
 
-    private boolean isSameOriginUsername(Long id, CustomerUpdateRequest request) {
+    private boolean isSameOriginUsername(long id, CustomerUpdateRequest request) {
         Customer foundCustomer = findById(id);
         return foundCustomer.isSameUsername(request.getUsername());
     }
 
-    public void delete(Long id) {
-        validateCustomerExists(id);
+    public void delete(long id, String password) {
+        validateDeletable(id, password);
         customerDao.delete(id);
     }
 
-    private void validateCustomerExists(Long id) {
-        findById(id);
+    private void validateDeletable(long id, String password) {
+        Customer customer = findById(id);
+        if (!customer.isSamePassword(password)) {
+            throw new InvalidCustomerException("비밀번호가 일치하지 않습니다.");
+        }
     }
 }
